test: reorganize shared components tests
